Move types to subdirectories
