+++ src/payload/issue/IssuePayload.ts
@@ -1,10 +1,10 @@
-import { Changes } from "./Changes";
+import { Changes } from "../Changes";
 import { IssueObjectAttributes } from "./IssueObjectAttributes";
-import { Label } from "./Label";
-import { Project } from "./Project";
-import { Repository } from "./Repository";
-import { User } from "./User";
-import { Payload } from "./Payload";
+import { Label } from "../Label";
+import { Project } from "../Project";
+import { Repository } from "../Repository";
+import { User } from "../User";
+import { Payload } from "../Payload";
 
 export interface IssuePayload extends Payload {
   object_kind: "issue";
