+++ src/payload/merge_request/MergeRequestObjectAttributes.ts
@@ -1,8 +1,8 @@
-import { Project } from "./Project";
-import { Author } from "./Author";
-import { ObjectAttributes } from "./ObjectAttributes";
-import { DateTime } from "./types";
-import { Commit } from "./Commit";
+import { Project } from "../Project";
+import { Author } from "../Author";
+import { ObjectAttributes } from "../ObjectAttributes";
+import { DateTime } from "../types";
+import { Commit } from "../Commit";
 
 export interface MergeRequestObjectAttributes extends ObjectAttributes {
   head_pipeline_id: number | null;
