+++ src/payload/merge_request/MergeRequestPayload.ts
@@ -1,10 +1,10 @@
-import { Changes } from "./Changes";
-import { Label } from "./Label";
+import { Changes } from "../Changes";
+import { Label } from "../Label";
 import { MergeRequestObjectAttributes } from "./MergeRequestObjectAttributes";
-import { Project } from "./Project";
-import { Repository } from "./Repository";
-import { User } from "./User";
-import { Payload } from "./Payload";
+import { Project } from "../Project";
+import { Repository } from "../Repository";
+import { User } from "../User";
+import { Payload } from "../Payload";
 
 export interface MergeRequestPayload extends Payload {
   object_kind: "merge_request";
