+++ src/payload/pipeline/PipelineObjectAttributes.ts
@@ -1,4 +1,4 @@
-import { DateTime } from "./types";
+import { DateTime } from "../types";
 import { Variables } from "./Variables";
 
 export interface PipelineObjectAttributes {
